rename component files to pascal case
