+++ client/src/styles/DashboardExerciseList.scss
