+++ client/src/styles/DashboardExerciseListItem.scss
