+++ client/src/styles/DayWorkoutList.scss
