+++ client/src/styles/DayWorkoutListItem.scss
